feat: SHOW TAG VALUES should produce results from one specific RP
Ensure that the Sources field of the ShowTagValuesStatement is
filled in. Then use the sources to limit the retention policies,
and thus the shards from which tag values are collected.

This fix only works on TSI databases; INMEM shards share
indices, so restricting shard indices used does not restrict
the tag values returned.

Closes https://github.com/influxdata/feature-requests/issues/265