fix: Increase loop length for making sure influx is up
With the current script, a large database doesn't have time to finish 
starting up in the 10 seconds alotted by this script.  This causes the 
service to fail to start up, and systemd to sit looping forever.

This PR increases the sleep value to 5 seconds, and the max_attempts to 
12, meaning the script will attempt to connect for up to 1 minute before 
giving up.

Part of #21967