feat: Enable freed memory tracking in heappy
In https://github.com/mkmik/heappy/commit/aed37ab50a70c3f0a7a8bd1c51d28d3d8050461f I fixed the main
problem that rendered free tracking useless: the allocator was actually allocating more bytes than requested; we were tracking the amount of memory requested by the user and not the size of the block that was returned to the user. However, when we were tracking a call to free we used the size of the memory block which was bigger than the amount that was tracked; this led to negative numbers of "in use memory".