refactor: re-export google proto types
The TemplatePart::Table() variant contains a google::protobuf::Empty
type which is defined in the generated_types crate.

Rather than force consumers to import two crates which have to be kept
in sync, this commit re-exports the google types within the
influxdb_iox_client crate as it's already a dependency, and makes using
types like protobuf's Duration or long-running-operation API easier for
the consumer.