fix(build): avoid BSD make evaluate Makefile
This Makefile use GNU make functions which is not compatible with the BSD
make. In order to avoid BSD make to try evaluation the Makefile and because
the Makefile use GNU make functions renaming the from Makefile to GNUmakefile
does not change anything for GNU make users and expose a clear error
when someone try to evaluate with BSD make.