build: feature-gate arrow dep for mgmt client
Avoid pulling in & compiling arrow for consumers of the managemnt
client.

PR #2098 added arrow_util as a dependency of this crate which in turn
pulls in all of arrow, however it is only used in the flight client.

This commit feature-gates arrow_util behind "flight" with the rest of
the arrow deps.