refactor: abstract HTTP inteface for different run modes
First iteration of a common interface for the different server run
modes. This splits the existing HTTP server into a part that can be used
by all run modes (e.g. health checks, metrics endpoint, profiling) and a
part that is specific to the the database run mode.

Note that the interface is not final and might require more iterations,
but we have to start somewhere.