fix: undected gzip HTTP body truncation
When reading the gzip-encoded body of a HTTP request, the stream is read
up until the configured maximum number of allowable bytes, at which
point the body was silently trucated. This could allow fields in
submitted line protocol to be silently lost (amongst other bad things).

This change ensures that truncation results in a RequestSizeExceeded
error.